LRQA-14421 split main.js into component.js and main.js, cleaned up code
