+++ modules/test/poshi-runner/src/META-INF/resources/html/index.html
@@ -7,6 +7,7 @@
 		<link href="http://maxcdn.bootstrapcdn.com/font-awesome/4.2.0/css/font-awesome.min.css" rel="stylesheet" />
 		<link href="../css/main.css" media="screen" rel="stylesheet" type="text/css" />
 		<script src="http://cdn.alloyui.com/3.0.1/aui/aui-min.js" type="text/javascript"></script>
+		<script src="../js/component.js" type="text/javascript"></script>
 		<script src="../js/main.js" type="text/javascript"></script>
 	</head>
 	<body>
